+++ Day-6-7-8-9/README.md
@@ -6,11 +6,11 @@ Todos
 - create a no result
 - refactoring the code
 
-### Day-6 assignment 
+### Day-6 assignment
 
 <img width="1440" alt="Screen Shot 2022-12-06 at 2 50 36 PM" src="https://user-images.githubusercontent.com/34391629/205871086-14752754-f0b5-4e56-b9f9-fbb1a3bf90b5.png">
 
-### Day-7 Assignment 
+### Day-7 Assignment
 
 - new route added
 
@@ -20,3 +20,5 @@ Todos
 
 <img width="769" alt="Screen Shot 2022-12-07 at 1 28 27 PM" src="https://user-images.githubusercontent.com/34391629/206121074-5f535772-1201-43d7-ba9a-8d04c2d0e5fe.png">
 
+- pages - for every route
+- components - fetching data different component
