+++ Day-6-7-8-9/src/components/UserRepoList.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { fetchUserData } from "../utils/fetchData";
-import Loader from "./Loader";
+import Loader from "./loader/Loader";
 import RepoCard from "./RepoCard";
 import WithRouter from "./WithRouter";
