+++ Day-6-7-8-9/src/redux/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import loginReducer from "./loginSlice";
+import loginReducer from "./slice/loginSlice";
 
 export const store = configureStore({
   reducer: {
